requirements.txt
pytelegrambotapi
python-dotenv
requests
schedule
pytz
